requirements.txt
flask==3.0.0
requests==2.31.0
line-bot-sdk==3.5.1
gunicorn==21.2.0
